backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
biopython==1.81
pandas==2.1.3
python-multipart==0.0.6
pydantic==2.5.0
